requirements.txt
python-dotenv
elevenlabs[pyaudio]
